requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=1.3.0=py310hecd8cb5_0
aiohttp=3.8.3=py310h6c40b1e_0
aiosignal=1.2.0=pyhd3eb1b0_0
anyio=3.5.0=py310hecd8cb5_0
appdirs=1.4.4=pypi_0
appnope=0.1.2=py310hecd8cb5_1001
argon2-cffi=20.1.0=py310hca72f7f_1
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=py310hecd8cb5_0
attrs=22.1.0=py310hecd8cb5_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py310hecd8cb5_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.4=py310hecd8cb5_0
bottleneck=1.3.5=py310h4e76f89_0
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
brotlipy=0.7.0=py310hca72f7f_1002
bzip2=1.0.8=h1de35cc_0
c-ares=1.18.1=hca72f7f_0
ca-certificates=2023.01.10=hecd8cb5_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.12.7=py310hecd8cb5_0
cffi=1.15.1=py310h6c40b1e_2
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py310hecd8cb5_0
comm=0.1.2=py310hecd8cb5_0
commonmark=0.9.1=pyhd3eb1b0_0
cryptography=38.0.1=py310hf6deb26_0
cycler=0.11.0=pyhd3eb1b0_0
data=0.4=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py310he9d5cce_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
docker-pycreds=0.4.0=pypi_0
entrypoints=0.4=py310hecd8cb5_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.3=h0a44026_0
fftw=3.3.9=h9ed2024_1
filelock=3.9.0=py310hecd8cb5_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=hd8bbffd_0
frozenlist=1.3.3=py310h6c40b1e_0
funcsigs=1.0.2=pypi_0
future=0.18.2=py310hecd8cb5_1
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
gitdb=4.0.10=pypi_0
gitpython=3.1.30=pypi_0
gmp=6.2.1=he9d5cce_3
gnutls=3.6.15=hed9c0bf_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
grpcio=1.42.0=py310ha29bfda_0
huggingface_hub=0.10.1=py310hecd8cb5_0
icu=58.2=h0a44026_3
idna=3.4=py310hecd8cb5_0
imageio=2.26.0=pyh24c5eb1_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.19.2=py310h20db666_0
ipython=8.7.0=py310hecd8cb5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.1=py310hecd8cb5_1
jinja2=3.1.2=py310hecd8cb5_0
joblib=1.0.1=pyh3f38642_3
jpeg=9e=hca72f7f_0
jsonschema=4.16.0=py310hecd8cb5_0
jupyter_client=7.4.8=py310hecd8cb5_0
jupyter_core=5.1.1=py310hecd8cb5_0
jupyter_server=1.23.4=py310hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
keopscore=2.1=pypi_0
kiwisolver=1.4.2=py310he9d5cce_0
lame=3.100=h1de35cc_0
latex=0.7.0=pypi_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libcxx=14.0.6=h9765a3e_0
libdeflate=1.8=h9ed2024_5
libffi=3.4.2=hecd8cb5_6
libgfortran=5.0.0=11_3_0_hecd8cb5_28
libgfortran5=11.3.0=h9dfd629_28
libiconv=1.16=hca72f7f_2
libidn2=2.3.2=h9ed2024_0
libpng=1.6.37=ha441bb4_0
libprotobuf=3.20.1=h8346a28_0
libsodium=1.0.18=h1de35cc_0
libtasn1=4.16.0=h9ed2024_0
libtiff=4.4.0=h2cd0358_2
libunistring=0.9.10=h9ed2024_0
libwebp=1.2.4=h56c3ce4_0
libwebp-base=1.2.4=hca72f7f_0
libxml2=2.9.14=hbf8cd5e_0
libxslt=1.1.35=h5b33f42_0
llvm-openmp=14.0.6=h0dcd299_0
lxml=4.9.1=py310h65b224f_0
lz4-c=1.9.3=h23ab428_1
markdown=3.4.1=py310hecd8cb5_0
markupsafe=2.1.1=py310hca72f7f_0
matplotlib=3.5.3=py310hecd8cb5_0
matplotlib-base=3.5.3=py310hfb0c5b7_0
matplotlib-inline=0.1.6=py310hecd8cb5_0
mistune=0.8.4=py310hca72f7f_1000
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py310hca72f7f_0
mkl_fft=1.3.1=py310hf879493_0
mkl_random=1.2.2=py310hc081a56_0
multidict=6.0.2=py310hca72f7f_0
munkres=1.1.4=py_0
nb_conda=2.2.1=py310hecd8cb5_1
nb_conda_kernels=2.3.1=py310hecd8cb5_0
nbclassic=0.4.8=py310hecd8cb5_0
nbclient=0.5.13=py310hecd8cb5_0
nbconvert=6.5.4=py310hecd8cb5_0
nbformat=5.7.0=py310hecd8cb5_0
ncurses=6.3=hca72f7f_3
nest-asyncio=1.5.6=py310hecd8cb5_0
nettle=3.7.3=h230ac6f_1
notebook=6.5.2=py310hecd8cb5_0
notebook-shim=0.2.2=py310hecd8cb5_0
numexpr=2.8.4=py310h9638375_0
numpy=1.23.4=py310h9638375_0
numpy-base=1.23.4=py310ha98c3c9_0
oauthlib=3.2.1=py310hecd8cb5_0
openh264=2.1.1=h8346a28_0
openssl=1.1.1t=hca72f7f_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.3=py310h3ea8b11_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py310hde71d04_1
pip=22.2.2=py310hecd8cb5_0
platformdirs=2.5.2=py310hecd8cb5_0
plotly=5.9.0=py310hecd8cb5_0
prometheus_client=0.14.1=py310hecd8cb5_0
prompt-toolkit=3.0.36=py310hecd8cb5_0
protobuf=3.20.1=py310he9d5cce_0
psutil=5.9.0=py310hca72f7f_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pybind11=2.10.1=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.4.0=py310hecd8cb5_0
pykeops=2.1=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py310hecd8cb5_0
pyrsistent=0.18.0=py310hca72f7f_0
pysocks=1.7.1=py310hecd8cb5_0
python=3.10.8=h218abb5_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py310hecd8cb5_0
pytorch=1.13.0=py3.10_0
pytz=2022.7=py310hecd8cb5_0
pyyaml=6.0=py310h6c40b1e_1
pyzmq=23.2.0=py310he9d5cce_0
readline=8.2=hca72f7f_0
regex=2022.7.9=py310hca72f7f_0
requests=2.28.1=py310hecd8cb5_0
requests-oauthlib=1.3.0=py_0
rich=12.5.1=py310hecd8cb5_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.1.1=py310he9d5cce_0
scipy=1.9.3=py310h09290a1_0
send2trash=1.8.0=pyhd3eb1b0_1
sentry-sdk=1.15.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=65.5.0=py310hecd8cb5_0
shutilwhich=1.1.0=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sniffio=1.2.0=py310hecd8cb5_1
soupsieve=2.3.2.post1=py310hecd8cb5_0
sqlite=3.40.0=h880c91c_0
stack_data=0.2.0=pyhd3eb1b0_0
tempdir=0.7.1=pypi_0
tenacity=8.0.1=py310hecd8cb5_1
tensorboard=2.10.0=py310hecd8cb5_0
tensorboard-data-server=0.6.1=py310h7242b5c_0
tensorboard-plugin-wit=1.6.0=py_0
terminado=0.17.1=py310hecd8cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=py310hecd8cb5_0
tk=8.6.12=h5d9f67b_0
tokenizers=0.11.4=py310h8776b5c_1
torchaudio=0.13.0=py310_cpu
torchvision=0.14.0=py310_cpu
tornado=6.2=py310hca72f7f_0
tqdm=4.64.1=py310hecd8cb5_0
traitlets=5.7.1=py310hecd8cb5_0
transformers=4.24.0=py310hecd8cb5_0
typing-extensions=4.3.0=py310hecd8cb5_0
typing_extensions=4.3.0=py310hecd8cb5_0
tzdata=2022f=h04d1e81_0
urllib3=1.26.12=py310hecd8cb5_0
wandb=0.13.10=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py310hecd8cb5_1
websocket-client=0.58.0=py310hecd8cb5_4
werkzeug=2.2.2=py310hecd8cb5_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.6=hca72f7f_0
yaml=0.2.5=haf1e3a3_0
yarl=1.8.1=py310hca72f7f_0
zeromq=4.3.4=h23ab428_0
zlib=1.2.13=h4dc903c_0
zstd=1.5.2=hcb37349_0
